-added sprite list loading -increased MAX_MGG to 256 -MGG mapping is now like this: the first 3 (0, 1, 2) are for UI, from slots 3 to 35, they're dedicated to the map's MGGs, from 36 to 255, they're all for sprites, HUD and stuff -use MGG_SPRITE_START macro to load the sprite's or HUD's MGGs -added MAX_MAPMGG, the maximum number of MGGs per map is 32 ( I believe that this is more than enough! ) -the first sprite list has 2 sprites defined: DUMMY and FULGORE, BOTH are for testing purposes -the sprite list is used ONLY for the mEngineer, since it doesn't have the actor's code

+++ mGear-1/mGear-1/engine.cpp
@@ -648,7 +648,7 @@ void FreeMGG(_MGG *file)
 
 void InitMGG()
 {
-	for(register uint8 i=0; i<MAX_MGG; i++)
+	for(register uint16 i=0; i<MAX_MGG; i++)
 	{
 		memset(&mgg[i],0,sizeof(_MGG));
 		mgg[i].type=NONE;
